refactor(examples-rest)!: drop standalone flag, restructure urls/ for ResolvedUrls
Switch `src/config/urls.rs` from `#[routes(standalone)]` to plain `#[routes]`.
The `standalone` flag suppresses `crate::urls::url_prelude` and the per-app
`ResolvedUrls::<app>()` accessor methods, and is documented as the right
choice only for projects that do *not* use `installed_apps!`. This example
does use `installed_apps!` (see `src/config/apps.rs`), so plain `#[routes]`
is the appropriate form — it generates the `ResolvedUrls::snippets()`
accessor backed by the resolver modules emitted by
`#[url_patterns(InstalledApp::snippets, mode = ...)]`.

Plain `#[routes]` requires every installed app to expose three resolver
modules under its `urls/` namespace:

- `crate::apps::<app>::urls::url_resolvers`            (mode = server)
- `crate::apps::<app>::urls::client_url_resolvers`     (mode = client)
- `crate::apps::<app>::urls::ws_urls::ws_url_resolvers` (mode = ws,
  hoisted under `urls/` in rc.19 per discussion #3914)

Restructure the `snippets` app to satisfy this:

- src/apps/snippets/urls.rs becomes a thin aggregator that declares
  `server_urls`, `client_router`, `ws_urls` submodules and re-exports the
  flat `url_resolvers` / `client_url_resolvers` modules at the `urls`
  level (the routes macro looks for the server and client resolvers at
  the flat path, but the WS one at the nested `ws_urls::ws_url_resolvers`
  path — that asymmetry is preserved).
- src/apps/snippets/urls/server_urls.rs takes over the existing
  `#[url_patterns(InstalledApp::snippets, mode = server)]` router (the
  function is now `server_url_patterns()` at the canonical path).
- src/apps/snippets/urls/client_router.rs is an empty
  `#[url_patterns(..., mode = client)] -> ClientRouter` stub. The
  REST-only example has no SPA, but the macro path must exist for
  `#[routes]` to compile.
- src/apps/snippets/urls/ws_urls.rs is an empty
  `#[url_patterns(..., mode = ws)] -> WebSocketRouter` stub.
- src/config/urls.rs calls `urls::server_urls::server_url_patterns()`.

Enable the `websockets` feature in Cargo.toml because `WebSocketRouter`
is only re-exported from `reinhardt` when that feature is on; the empty
stub still requires the type to be in scope. The `websockets` feature
pulls in `tokio-tungstenite`, which is the unavoidable cost of using
non-`standalone` `#[routes]` in an app that has no real WS surface.

Verification (all from examples/examples-tutorial-rest):
- cargo check --all-features                 → clean
- cargo nextest run --all-features            → 26/26 passing
- cargo fmt --check                           → clean
- cargo clippy --all-features --all-targets -- -D warnings → clean
